requirements.txt

streamlit
torch
torchvision
opencv-python
pillow
pyyaml
numpy
matplotlib
